requirements/requirements.txt
dj-database-url==0.4.1
git+git://github.com/django-admin-bootstrapped/django-admin-bootstrapped.git
django-bootstrap3==7.0.1
django-bower==5.1.0
django-braces==1.8.1
django-compressor==2.0
django-crispy-forms==1.6.0
django-extensions==1.6.1
django-extra-views==0.7.1
django-filter==0.13.0
django-floppyforms==1.6.1
django-model-utils==2.4
django-sass-processor==0.3.4
django-image-cropping==1.0.3
django==1.9.5
djangorestframework
easy-thumbnails==2.3
libsass==0.11.0
psycopg2==2.6.1
python-decouple==3.0
pytz==2016.3
pyyaml==3.11
rtyaml==0.0.2
unipath==1.1
python-magic==0.4.10
gunicorn==19.4.5
python-social-auth
